+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- [\#xx](https://github.com/medibloc/panacea-core/pull/xx) Add a new feature for issuing new currencies (by a new module `x/currency`)
+- [\#80](https://github.com/medibloc/panacea-core/pull/80) Add the `x/token` module for issuing new tokens
 
 ## [v1.3.2](https://github.com/medibloc/panacea-core/releases/tag/v1.3.2) - 2020-11-04